use size_t

+++ src/Particle/SampleStack.h
@@ -31,20 +31,20 @@ class SampleStack
 public:
   using PropertySetType = QMCTraits::PropertySetType;
 
-  void setTotalNum(int total_num) { total_num_ = total_num; }
+  void setTotalNum(size_t total_num) { total_num_ = total_num; }
 
-  int getMaxSamples() const { return max_samples_; }
+  size_t getMaxSamples() const { return max_samples_; }
 
   bool empty() const { return sample_vector_.empty(); }
 
   const MCSample& getSample(size_t i) const;
 
   //@{save/load/clear function for optimization
-  inline int getNumSamples() const { return current_sample_count_; }
+  inline size_t getNumSamples() const { return current_sample_count_; }
   ///set the number of max samples per rank.
-  void setMaxSamples(int n, int number_of_ranks = 1);
+  void setMaxSamples(size_t n, size_t number_of_ranks = 1);
   /// Global number of samples is number of samples per rank * number of ranks
-  int64_t getGlobalNumSamples() const { return global_num_samples_; }
+  size_t getGlobalNumSamples() const { return global_num_samples_; }
   /// load a single sample from SampleStack
   void loadSample(ParticleSet& pset, size_t iw) const;
 
@@ -57,10 +57,10 @@ public:
   void resetSampleCount();
 
 private:
-  int total_num_{0};
-  int max_samples_{10};
-  int current_sample_count_{0};
-  int64_t global_num_samples_{max_samples_};
+  size_t total_num_{0};
+  size_t max_samples_{10};
+  size_t current_sample_count_{0};
+  size_t global_num_samples_{max_samples_};
 
   std::vector<MCSample> sample_vector_;
 };
